+++ src/pages/DailyCheck/store/index.ts
